+++ CHANGES.md
@@ -5,7 +5,7 @@
 * Module `xcube.core.mldataset` has been refactored into 
   a sub-package for clarity and maintainability.
 
-* Provided backward compatibility with Python 3.8. (#)
+* Provided backward compatibility with Python 3.8. (#760)
 
 ## Changes in 0.13.0.dev6